refactor(aicore/filtering): merge _litellm_patch + core/env into filters.py
Address review comments #17 and #18 (part 1):

- Merge _litellm_patch.py (LiteLLM transport patch, _install, _active_cfg,
  _ORIGINAL_CONFIG, FilteringOrchestrationConfig, extract_filter_blocked)
  into filters.py. The whole filtering surface (Severity, ContentFilter,
  AzureContentFilter, LlamaGuard38bFilter, InputFiltering, OutputFiltering,
  ContentFiltering, set_filtering, disable_filtering, extract_filter_blocked,
  FilteringOrchestrationConfig) now lives in one file at ~565 lines.
- Inline _read_env_str / _read_env_bool / _read_env_choice helpers from
  core/env.py into filters.py (private module-level helpers). Delete
  src/sap_cloud_sdk/core/env.py + tests/core/unit/test_env.py — single
  consumer, no need for a shared module yet.

aicore/filtering/__init__.py stays a thin re-export of the public names
plus the exception types. exceptions.py is the only other file in the
package.

Test imports updated: test_filtering.py and test_patch.py now import
_install / _ORIGINAL_CONFIG / FilteringOrchestrationConfig from
.filters (not ._litellm_patch). Mock paths in test_patch.py updated to
match the new module location.
